[rs] Refactor to prepare `tag` fuzzing
