+++ rs/Cargo.toml
@@ -29,7 +29,7 @@ nom = "^5.0.0"
 num-traits = "^0.2.8"
 regex = "^1.1.8"
 serde_json = "^1.0.40"
-swf-tree = "^0.9.0"
+swf-tree = { git="https://github.com/open-flash/swf-tree.git", rev = "524f19adc098d83900e94bbe693887a63b50786d" }
 swf-fixed = "^0.1.4"
 
 [dev-dependencies]
@@ -37,8 +37,8 @@ serde = "^1.0.94"
 serde_json_v8 = "^0.0.1"
 test-generator = "^0.2.2"
 
-[replace]
-"swf-tree:0.9.0" = { git="https://github.com/open-flash/swf-tree.git", rev = "524f19adc098d83900e94bbe693887a63b50786d" }
+# [replace]
+# "swf-tree:0.9.0" = { git="https://github.com/open-flash/swf-tree.git", rev = "524f19adc098d83900e94bbe693887a63b50786d" }
 
 # When testing larger files, increasing `opt-level` provides a significant speed-up.
 # [profile.test]
