+++ rs/fuzz/Cargo.toml
@@ -3,14 +3,14 @@ name = "swf-parser-fuzz"
 version = "0.0.1"
 authors = ["Automatically generated"]
 publish = false
+edition = "2018"
 
 [package.metadata]
 cargo-fuzz = true
 
-[dependencies.swf-parser]
-path = ".."
-[dependencies.libfuzzer-sys]
-git = "https://github.com/rust-fuzz/libfuzzer-sys.git"
+[dependencies]
+swf-parser = {path = ".."}
+libfuzzer-sys = "^0.1.1"
 
 # Prevent this from interfering with workspaces
 [workspace]
@@ -19,3 +19,7 @@ members = ["."]
 [[bin]]
 name = "swf"
 path = "fuzz_targets/swf.rs"
+
+[[bin]]
+name = "tag"
+path = "fuzz_targets/tag.rs"
