+++ rs/fuzz/fuzz_targets/swf.rs
@@ -1,6 +1,6 @@
 #![no_main]
-#[macro_use] extern crate libfuzzer_sys;
-extern crate swf_parser;
+
+use libfuzzer_sys::fuzz_target;
 
 fuzz_target!(|data: &[u8]| {
    let _ = swf_parser::complete::parse_swf(data);
