+++ rs/src/lib.rs
@@ -8,15 +8,10 @@ pub use swf_tree;
 mod tests {
   use std::io::{Read, Write};
   use std::path::Path;
-
   use ::swf_tree::Movie;
   use nom::IResult as NomResult;
-  use swf_tree::Tag;
-
   use ::test_generator::test_expand_paths;
-
   use crate::complete::parse_swf;
-  use crate::complete::parse_tag;
 
   test_expand_paths! { test_parse_movie; "../tests/movies/*/" }
   fn test_parse_movie(path: &str) {
@@ -53,35 +48,6 @@ mod tests {
     let expected_movie = serde_json_v8::from_reader::<_, Movie>(ast_reader).expect("Failed to read AST");
 
     assert_eq!(actual_movie, expected_movie);
-  }
-
-  test_expand_paths! { test_parse_tag; "../tests/tags/*/*/" }
-  fn test_parse_tag(path: &str) {
-    let path: &Path = Path::new(path);
-    let name = path
-      .components()
-      .last()
-      .unwrap()
-      .as_os_str()
-      .to_str()
-      .expect("Failed to retrieve sample name");
-    let input_path = path.join("input.bytes");
-    let input_bytes: Vec<u8> = ::std::fs::read(input_path).expect("Failed to read input");
-
-    let swf_version: u8 = match name {
-      "po2-swf5" => 5,
-      _ => 10,
-    };
-
-    let (remaining_bytes, actual_value) = parse_tag(&input_bytes, swf_version);
-
-    let expected_path = path.join("value.json");
-    let expected_file = ::std::fs::File::open(expected_path).expect("Failed to open expected value file");
-    let expected_reader = ::std::io::BufReader::new(expected_file);
-    let expected_value = serde_json_v8::from_reader::<_, Tag>(expected_reader).expect("Failed to read AST");
-
-    assert_eq!(actual_value, Some(expected_value));
-    assert_eq!(remaining_bytes, &[] as &[u8]);
   }
 
   macro_rules! test_various_parser_impl {
